Resolved a conflict in the add service consumtion -timaoffManagment

+++ FrontEnd/CareerLink1/src/app/components/timeofftracker/timeofftracker.component.ts
@@ -1,8 +1,12 @@
 import { Component } from '@angular/core';
 import { FormBuilder, FormGroup ,Validators} from '@angular/forms';
 import { Observable } from 'rxjs';
+import { LeaveStatus } from 'src/app/models/LeaveStatus';
 import { TimeOffTracker } from 'src/app/models/TimeOffTracker';
+import { User } from 'src/app/models/User';
 import { TimeofftrackerService } from 'src/app/services/timeofftracker.service';
+import { UserService } from 'src/app/services/user.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-timeofftracker',
@@ -16,19 +20,28 @@ export class TimeofftrackerComponent {
    timeoffForm ! :FormGroup;
    leaveType :String[]=['Casual','Compassionate','Medical','Maternity','Other'];
    leaveStatus :String[]=['Pending','Accepted','Rejected'];
-
+   users: User[] = [];
 
    ngOnInit() :void {
      this.LoadListOfTimesOf();
+     
      this.createForm();
+     this.loadUsers();
+   
    }
 
-   constructor(private timeoffService :TimeofftrackerService , private formbilder: FormBuilder) {   
-  }
+   constructor(private timeoffService :TimeofftrackerService , private formbilder: FormBuilder, private userService: UserService) { }
+  
+   LoadListOfTimesOf() : void{ 
+
+      this.timeoffService.findAllTimesOff().subscribe( (timesOff:TimeOffTracker[] )=> {
+        this.timesOff=  timesOff;
+      } );
+   }
    createForm() :void{
     this.timeoffForm = this.formbilder.group({
     
-      leaveType: ['', Validators.required],
+      type: ['', Validators.required],
       description : ['', Validators.required],
       fromDate : ['', Validators.required],
       toDate : ['', Validators.required],
@@ -36,14 +49,11 @@ export class TimeofftrackerComponent {
       
     });
    }
-   LoadListOfTimesOf() : void{  //Always add subscribe because observables are lazy 
-      this.timeoffService.findAllTimesOff().subscribe( (timesOff:TimeOffTracker[] )=> {
-        this.timesOff=  timesOff;
-      } );
-   }
-
    addTimeOff(): void {
+    
     const newtimeoff = this.timeoffForm.value;
+   
+
     this.timeoffService.TakeTiMEOff(newtimeoff)
     .subscribe(
       response => {
@@ -52,17 +62,45 @@ export class TimeofftrackerComponent {
       },
       error => console.error('error, add', error)
     );
-  }
-  
- 
+  } 
+
+
  deletetimeOff(id: number):void {
-  this.timeoffService.deleteTiMEOff(id).subscribe ( ():void => {
-  this.LoadListOfTimesOf();
+  Swal.fire({
+    title: "Are you sure?",
+    text: "You won't be able to revert this!",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Yes, delete it!"
+  }).then((result) => {
+    if (result.isConfirmed) {
+      this.timeoffService.deleteTiMEOff(id).subscribe ( ():void => {
+    
+        this.LoadListOfTimesOf();
+        });
+      Swal.fire({
+        title: "Deleted!",
+        text: "this time Off rquest has been deleted.",
+        icon: "success"
+      });
+    }
   });
+  
   }
  cancel():void {
  this.timeoffForm.reset();
  }
+ loadUsers(): void{
+  this.userService.findAllUsers()
+  .subscribe(
+    users => this.users = users,
+    error => console.error('error, getall', error)
+  );
+}
+
+
 
 }
